Refactor OAuth2AuthHandlerImpl
