+++ vertx-web-auth-common/src/main/java/io/vertx/ext/auth/common/AuthenticationContext.java
@@ -31,6 +31,8 @@ public interface AuthenticationContext {
    */
   UserContext user();
 
+  String normalizedPath();
+
   default void onContinue() {
     // NOOP
   }
