+++ vertx-web-auth-common/src/main/java/io/vertx/ext/auth/common/AuthenticationContextInternal.java
@@ -1,0 +1,17 @@
+package io.vertx.ext.auth.common;
+
+import io.vertx.ext.auth.audit.SecurityAudit;
+
+public interface AuthenticationContextInternal {
+
+  /**
+   * Get or Default the security audit object.
+   */
+  SecurityAudit securityAudit();
+
+  /**
+   * Get or Default the security audit object.
+   */
+  void setSecurityAudit(SecurityAudit securityAudit);
+
+}
